+++ crates/miden-lib/src/testing/mock_util_lib.rs
@@ -5,7 +5,7 @@ use miden_objects::utils::sync::LazyLock;
 use crate::transaction::TransactionKernel;
 
 const MOCK_UTIL_LIBRARY_CODE: &str = "
-    use.miden::tx
+    use.miden::output_note
 
     # Inputs:  []
     # Outputs: [note_idx]
@@ -17,7 +17,7 @@ const MOCK_UTIL_LIBRARY_CODE: &str = "
         push.0xc0000000        # = NoteTag::LocalAny
         # => [tag, aux, note_type, execution_hint, RECIPIENT, pad(8)]
 
-        exec.tx::create_note
+        exec.output_note::create
         # => [note_idx]
     end
 
@@ -30,7 +30,7 @@ const MOCK_UTIL_LIBRARY_CODE: &str = "
         movdn.4
         # => [ASSET, note_idx]
 
-        exec.tx::add_asset_to_note dropw
+        exec.output_note::add_asset dropw
         # => [note_idx]
     end
 ";