+++ crates/miden-testing/tests/scripts/p2id.rs
@@ -219,14 +219,14 @@ fn test_create_consume_multiple_notes() -> anyhow::Result<()> {
 
     let tx_script_src = &format!(
         "
-            use.miden::tx
+            use.miden::output_note
             begin
                 push.{recipient_1}
                 push.{note_execution_hint_1}
                 push.{note_type_1}
                 push.0              # aux
                 push.{tag_1}
-                call.tx::create_note
+                call.output_note::create
 
                 push.{asset_1}
                 call.::miden::contracts::wallets::basic::move_asset_to_note
@@ -237,7 +237,7 @@ fn test_create_consume_multiple_notes() -> anyhow::Result<()> {
                 push.{note_type_2}
                 push.0              # aux
                 push.{tag_2}
-                call.tx::create_note
+                call.output_note::create
 
                 push.{asset_2}
                 call.::miden::contracts::wallets::basic::move_asset_to_note